refactor: types entities model로 이동
