+++ src/entities/post/model/types.ts
