+++ src/entities/tag/model/index.ts
@@ -1,6 +1,6 @@
 import { useEffect, useState } from "react"
 import { fetchTagsApi } from "../api"
-import { Tag } from "../types"
+import { Tag } from "./types"
 
 export const useTags = () => {
   const [tags, setTags] = useState<Tag[]>([])
