+++ src/features/comment-item/ui/CommentItem.tsx
@@ -2,7 +2,7 @@ import { Edit2, ThumbsUp, Trash2 } from "lucide-react"
 import { Button } from "../../../shared/ui"
 import { highlightText } from "../../../shared/lib/highlightText"
 import { useCommentContext } from "../../../entities/comment/model/CommentContext"
-import { Comment } from "../../../entities/comment/types"
+import { Comment } from "../../../entities/comment/model/types"
 import { usePostParamsContext } from "../../../entities/post/model/PostParamsContext"
 
 interface Props {
