+++ src/features/post-view/model/usePosts.ts
@@ -1,7 +1,7 @@
 import { useState } from "react"
 import { deletePostApi, fetchPostsApi } from "../../../entities/post/api"
 import { fetchUsersApi } from "../../../entities/user/api"
-import { Post } from "../../../entities/post/types"
+import { Post } from "../../../entities/post/model/types"
 import { User } from "../../../entities/user/model/types"
 import { fetchPostsByTagApi, searchPostsApi } from "../api"
